+++ back/src/main/java/com/est/back/blindChat/dto/FeedbackDto.java
@@ -1,0 +1,4 @@
+package com.est.back.blindChat.dto;
+
+public class FeedbackDto {
+}
